feat: use external data and question
do not generate the data inside the component it will come from the
outside
