+++ public/index.html
@@ -27,6 +27,7 @@
 	<div class="nav">
 		<button id="prev">Prev</button>
 		<button id="next">Next</button>
+		<button id="restart">Restart</button>
 	</div>
 </body>
 </html>
